Task 1

+++ 0x1E-search_algorithms/1-binary.c
@@ -9,7 +9,7 @@
  */
 int binary_search(int *array, size_t size, int value)
 {
-	size_t i, left, right;
+	int i, left, right;
 
 	if (array == NULL)
 		return (-1);
